Use some global njit functions in numba
This allows numba to reuse previous typing and compilation
results if the same function is reused, which then also
leads to smaller llvm modules.

For the tests to continue to work we have to return
those global functions through a wrapper (`basic.global_numba_func`)
so that the tests are still able to disable compilation.

Also remove some inline="always" arguments that don't seem
to be helpful.