Remove `until_condition_failed` in ScalarLoop
This was problematic when OpenMP was used in the Elemwise outer loop

We add one extra output flag stating whether iteration converged or not. This however breaks Hyp2F1 grad in python mode because it goes beyond the Elemwise limit on number of operands. To fix it we split the grad when on python mode