Explicitly prevent Python 3.12
To discuss: If we don't support 3.12, maybe we should enforce this here in `pyproject.toml`?

Source: <https://github.com/pymc-devs/pytensor/issues/441#issuecomment-1752718880>

I'm not convinced that this is the right move since it might slightly hinder testing. But if we are explicit, then this seems to me like the correct place for it.