Added macros for Numpy 1.x compatibility
These need to be copied into the source code
since they're not available in Numpy 1.x.

They define macros with the Numpy 1.x behavior
for getting real and imaginary parts. In Numpy 2.0
these macros are already defined with the new definitions,
so this addition to our code will be ignored in Numpy 2.0