Split up TestMinMax::test_uint
I split this test up to test uint64 separately, since
this is the case discussed in Issue #770. I also added
a test for the exact example used in that issue.

The uint dtypes with lower precision should pass.

The uint64 case started passing for me locally on Mac OSX,
but still fails on CI. I'm not sure why this is, but at
least the test will be more specific now if it fails in the future.