Fix bug in handling of row/column matrices in GEMV c_code
Bug was caused by reusing the adjusted strides in the logic to decide whether the call to GEMV should be transposed or not.

Particularly the +1 in the strides variable was causing the error branch (no double-strides) to be reached wrongly. The +1 was supposedly there for the case of matrix with length 0, but that triggers a branch where the adjusted strides are never used.

This bug was introduced in afe934b229f15be4dcd96bbd3960e7ed328d99fa